API: add support for optional dependency charger_info
The current implementation allows to load general charger
information from a YAML file which is then provided via MQTT.

The newer ChargerInfo module can provide the same information.
This change adapts the API module so that it honours it.

To allow backward compatibility, the functionlity to load
the data from the file is still left. However, it is not
used (the loaded data is replaced) when the optional
dependency is used.

Note: that the current keys in the YAML file differ from
the new ChargerInfo interface. It is recommend to switch
to the newer interface to have a uniform presentation of
EVerest powered charging stations.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>